add: feature:authモジュールの追加
