+++ feature/auth/build.gradle.kts
@@ -1,34 +1,24 @@
 plugins {
-    alias(libs.plugins.android.application)
+    alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
-    kotlin("kapt")
-    alias(libs.plugins.hilt)
 }
 
 android {
-    namespace = "com.example.taskmate"
+    namespace = "com.example.feature"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.example.taskmate"
         minSdk = 26
-        targetSdk = 34
-        versionCode = 1
-        versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        vectorDrawables {
-            useSupportLibrary = true
-        }
     }
 
     buildTypes {
         release {
             isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
+        }
+        debug {
+            isMinifyEnabled = false
         }
     }
     compileOptions {
@@ -61,27 +51,12 @@ dependencies {
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.material3)
     implementation(libs.androidx.navigation.compose)
-    implementation(libs.hilt.android)
-    implementation(project(":core:data"))
-    implementation(project(":core:ui"))
-    implementation(project(":feature:home"))
-    implementation(project(":feature:task"))
-    implementation(project(":feature:mypage"))
-    implementation(project(":feature:auth"))
-    implementation(project(":feature:setting"))
-    kapt(libs.hilt.compiler)
+    implementation(libs.androidx.appcompat)
+    implementation(libs.material)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
-    androidTestImplementation(platform(libs.androidx.compose.bom))
-    androidTestImplementation(libs.androidx.ui.test.junit4)
-    debugImplementation(libs.androidx.ui.tooling)
-    debugImplementation(libs.androidx.ui.test.manifest)
-}
 
-kapt {
-    correctErrorTypes = true
-    arguments {
-        arg("dagger.fastInit", "ENABLED") // Daggerの初期化プロセスを高速化
-    }
+    implementation(project(":core:data"))
+    implementation(project(":core:ui"))
 }
