+++ feature/auth/src/main/java/com/example/feature/auth/FirstAuthScreen.kt
@@ -1,4 +1,4 @@
-package com.example.taskmate.ui.auth
+package com.example.feature.auth
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
