+++ feature/auth/src/main/res/drawable/account.xml
@@ -1,0 +1,20 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="25dp"
+    android:height="25dp"
+    android:viewportWidth="25"
+    android:viewportHeight="25">
+  <path
+      android:pathData="M12.5,5.5C14.16,5.5 15.5,6.84 15.5,8.5C15.5,10.16 14.16,11.5 12.5,11.5C10.84,11.5 9.5,10.16 9.5,8.5C9.5,6.84 10.84,5.5 12.5,5.5Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#545F70"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M12.5,14.5C16.5,14.5 19.5,16.5 19.5,17.5V19.5H5.5V17.5C5.5,16.5 8.5,14.5 12.5,14.5Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#545F70"
+      android:strokeLineCap="round"/>
+</vector>
