+++ feature/auth/src/main/res/drawable/book.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="18dp"
+    android:height="20dp"
+    android:viewportWidth="18"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M17.75,14.714C17.749,14.762 17.744,14.81 17.736,14.857C17.757,14.966 17.754,15.077 17.727,15.185C17.699,15.292 17.649,15.392 17.579,15.477C17.508,15.562 17.42,15.631 17.32,15.678C17.22,15.726 17.111,15.75 17,15.75H3C2.836,15.75 2.673,15.782 2.522,15.845C2.37,15.908 2.232,16 2.116,16.116C2,16.232 1.908,16.37 1.845,16.522C1.782,16.673 1.75,16.836 1.75,17C1.75,17.164 1.782,17.327 1.845,17.478C1.908,17.63 2,17.768 2.116,17.884C2.232,18 2.37,18.092 2.522,18.155C2.673,18.218 2.836,18.25 3,18.25H17C17.199,18.25 17.39,18.329 17.53,18.47C17.671,18.61 17.75,18.801 17.75,19C17.75,19.199 17.671,19.39 17.53,19.53C17.39,19.671 17.199,19.75 17,19.75H3C2.271,19.75 1.571,19.46 1.055,18.944C0.54,18.429 0.25,17.729 0.25,17V3C0.25,2.271 0.54,1.571 1.055,1.055C1.571,0.54 2.271,0.25 3,0.25H16.4C17.146,0.25 17.75,0.854 17.75,1.6V14.714ZM6,4.25C5.801,4.25 5.61,4.329 5.47,4.47C5.329,4.61 5.25,4.801 5.25,5C5.25,5.199 5.329,5.39 5.47,5.53C5.61,5.671 5.801,5.75 6,5.75H12C12.199,5.75 12.39,5.671 12.53,5.53C12.671,5.39 12.75,5.199 12.75,5C12.75,4.801 12.671,4.61 12.53,4.47C12.39,4.329 12.199,4.25 12,4.25H6Z"
+      android:fillColor="#545F70"
+      android:fillType="evenOdd"/>
+</vector>
