+++ feature/auth/src/main/res/drawable/setting.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="25dp"
+    android:height="25dp"
+    android:viewportWidth="25"
+    android:viewportHeight="25">
+  <path
+      android:pathData="M22.957,13.325C22.753,13.097 22.641,12.804 22.641,12.5C22.641,12.196 22.753,11.903 22.957,11.675L24.586,9.875C24.765,9.679 24.877,9.431 24.904,9.169C24.932,8.906 24.874,8.641 24.739,8.413L22.194,4.089C22.06,3.861 21.857,3.68 21.612,3.572C21.368,3.464 21.095,3.435 20.833,3.489L18.441,3.964C18.136,4.025 17.819,3.976 17.55,3.824C17.28,3.672 17.077,3.428 16.977,3.139L16.201,0.852C16.116,0.603 15.953,0.388 15.736,0.235C15.519,0.083 15.259,0.001 14.993,0.002H9.903C9.626,-0.013 9.351,0.063 9.121,0.216C8.892,0.37 8.719,0.593 8.631,0.852L7.918,3.139C7.819,3.428 7.615,3.672 7.346,3.824C7.076,3.976 6.76,4.025 6.455,3.964L4,3.489C3.751,3.454 3.497,3.493 3.271,3.6C3.045,3.706 2.855,3.876 2.727,4.089L0.183,8.413C0.044,8.639 -0.018,8.902 0.005,9.164C0.027,9.427 0.134,9.676 0.31,9.875L1.926,11.675C2.13,11.903 2.242,12.196 2.242,12.5C2.242,12.804 2.13,13.097 1.926,13.325L0.31,15.125C0.134,15.324 0.027,15.573 0.005,15.835C-0.018,16.098 0.044,16.361 0.183,16.587L2.727,20.911C2.861,21.139 3.065,21.32 3.309,21.428C3.553,21.536 3.826,21.565 4.089,21.511L6.481,21.036C6.785,20.975 7.102,21.024 7.371,21.176C7.641,21.328 7.845,21.572 7.944,21.861L8.72,24.148C8.808,24.407 8.981,24.63 9.21,24.784C9.44,24.937 9.715,25.013 9.992,24.998H15.082C15.348,24.999 15.608,24.917 15.825,24.765C16.042,24.612 16.205,24.397 16.29,24.148L17.066,21.861C17.166,21.572 17.369,21.328 17.639,21.176C17.908,21.024 18.225,20.975 18.53,21.036L20.922,21.511C21.184,21.565 21.457,21.536 21.701,21.428C21.946,21.32 22.149,21.139 22.283,20.911L24.828,16.587C24.963,16.359 25.021,16.094 24.993,15.831C24.966,15.569 24.854,15.321 24.675,15.125L22.957,13.325ZM21.062,15L22.08,16.125L20.451,18.899L18.949,18.599C18.033,18.415 17.08,18.568 16.271,19.029C15.462,19.49 14.853,20.226 14.56,21.099L14.076,22.499H10.819L10.361,21.074C10.068,20.201 9.46,19.465 8.65,19.004C7.841,18.543 6.888,18.39 5.972,18.574L4.47,18.874L2.816,16.112L3.834,14.987C4.46,14.3 4.806,13.41 4.806,12.488C4.806,11.565 4.46,10.675 3.834,9.988L2.816,8.863L4.445,6.113L5.946,6.413C6.863,6.597 7.816,6.444 8.625,5.984C9.434,5.523 10.043,4.786 10.336,3.914L10.819,2.501H14.076L14.56,3.926C14.853,4.799 15.462,5.535 16.271,5.996C17.08,6.457 18.033,6.61 18.949,6.426L20.451,6.126L22.08,8.9L21.062,10.025C20.443,10.711 20.101,11.596 20.101,12.512C20.101,13.429 20.443,14.314 21.062,15ZM12.448,7.501C11.441,7.501 10.457,7.794 9.62,8.343C8.783,8.893 8.131,9.673 7.746,10.587C7.361,11.5 7.26,12.505 7.456,13.475C7.653,14.445 8.137,15.336 8.849,16.035C9.561,16.734 10.468,17.21 11.455,17.403C12.442,17.596 13.465,17.497 14.396,17.119C15.325,16.74 16.12,16.1 16.68,15.278C17.239,14.455 17.537,13.489 17.537,12.5C17.537,11.174 17.001,9.903 16.047,8.965C15.092,8.027 13.798,7.501 12.448,7.501ZM12.448,15C11.945,15 11.453,14.853 11.034,14.578C10.616,14.304 10.29,13.913 10.097,13.457C9.904,13 9.854,12.497 9.952,12.012C10.05,11.528 10.293,11.082 10.649,10.733C11.004,10.383 11.458,10.145 11.951,10.048C12.445,9.952 12.957,10.001 13.422,10.191C13.887,10.38 14.284,10.7 14.564,11.111C14.843,11.522 14.993,12.006 14.993,12.5C14.993,13.163 14.724,13.799 14.247,14.267C13.77,14.736 13.123,15 12.448,15Z"
+      android:fillColor="#545F70"/>
+</vector>
