+++ feature/auth/src/main/res/drawable/task.xml
@@ -1,0 +1,20 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="26dp"
+    android:height="25dp"
+    android:viewportWidth="26"
+    android:viewportHeight="25">
+  <path
+      android:pathData="M4.5,9.868C4.5,6.395 4.5,4.658 5.525,3.579C6.55,2.5 8.2,2.5 11.5,2.5H14.5C17.8,2.5 19.45,2.5 20.475,3.58C21.5,4.657 21.5,6.394 21.5,9.867V15.131C21.5,18.604 21.5,20.341 20.475,21.42C19.45,22.499 17.8,22.5 14.5,22.5H11.5C8.2,22.5 6.55,22.5 5.525,21.42C4.5,20.343 4.5,18.606 4.5,15.133V9.868Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#545F70"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M9,2.5L9.082,2.993C9.282,4.19 9.382,4.789 9.802,5.145C10.22,5.5 10.827,5.5 12.041,5.5H13.958C15.171,5.5 15.778,5.5 16.198,5.145C16.618,4.789 16.718,4.19 16.917,2.993L17,2.5M9,16.5H13M9,11.5H17"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#545F70"
+      android:strokeLineCap="round"/>
+</vector>
