+++ settings.gradle.kts
@@ -24,6 +24,7 @@ include(":app")
 include(":feature:home")
 include(":feature:task")
 include(":feature:mypage")
+include(":feature:auth")
 include(":feature:setting")
 include(":core:data")
 include(":core:ui")
